pyproject.toml
[project]
name = "algorithmic-trading-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fake-useragent>=2.2.0",
    "fear-and-greed>=0.4",
    "jupyter>=1.1.1",
    "jupyter-client>=8.6.3",
    "jupyter-console>=6.6.3",
    "jupyter-core>=5.7.2",
    "latex>=0.7.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pandas-datareader>=0.10.0",
    "polygon-api-client>=1.14.5",
    "requests>=2.32.3",
    "ruff>=0.11.10",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.2.0",
    "statsmodels>=0.14.4",
    "tabulate>=0.9.0",
    "xlsxwriter>=3.2.3",
    "yfinance>=0.2.61",
]
